Move `CreateDatadogBrowserRouter` in the `react-router` folder
